+++ xc3_model/src/lib.rs
@@ -62,7 +62,7 @@ pub use material::{
 pub use sampler::{AddressMode, FilterMode, Sampler};
 pub use skeleton::{Bone, Skeleton};
 pub use texture::{ImageFormat, ImageTexture, ViewDimension};
-pub use xc3_lib::mxmd::{BlendState, CullMode, RenderPassType, StateFlags};
+pub use xc3_lib::mxmd::{BlendMode, CullMode, RenderPassType, StateFlags};
 
 pub mod animation;
 pub mod gltf;